[Update] Split MyProfile.svelte into several files with the different events, added feed on my profile
